add user controller, user auth, change db relationship

+++ src/main/java/project/library/controllers/BookController.java
@@ -1,5 +1,6 @@
 package project.library.controllers;
 
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 import project.library.dtos.BookRecordDTO;
 import project.library.models.Book;
 import project.library.repositories.BookRepository;
+import project.library.repositories.UserRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,9 +21,17 @@ public class BookController {
 
     @Autowired
     BookRepository bookRepository;
+    @Autowired
+    UserRepository userRepository;
 
     @PostMapping("/books")
-    public ResponseEntity<Book> createBook(@RequestBody @Valid BookRecordDTO bookRecordDTO) {
+    public ResponseEntity<Object> createBook(@RequestBody @Valid BookRecordDTO bookRecordDTO, HttpServletRequest request) {
+        var user = userRepository.findById((UUID) request.getAttribute("idUser"));
+
+        if (user.get().getAdmin().equals(false)) {
+            return ResponseEntity.status(HttpStatus.CREATED).body("Not allowed");
+        }
+
         var book = new Book();
         BeanUtils.copyProperties(bookRecordDTO, book);
         return ResponseEntity.status(HttpStatus.CREATED).body(bookRepository.save(book));
@@ -48,7 +58,14 @@ public class BookController {
     }
 
     @PutMapping("/books/{id}")
-    public ResponseEntity<Object> updateBook(@PathVariable(value = "id") UUID id, @RequestBody @Valid BookRecordDTO bookRecordDTO) {
+    public ResponseEntity<Object> updateBook(@PathVariable(value = "id") UUID id,
+                                             @RequestBody @Valid BookRecordDTO bookRecordDTO, HttpServletRequest request) {
+
+        var user = userRepository.findById((UUID) request.getAttribute("idUser"));
+        if(user.get().getAdmin().equals(false)) {
+            return ResponseEntity.status(HttpStatus.CREATED).body("Not allowed");
+        }
+
         Optional<Book> book = bookRepository.findById(id);
         if(book.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Book not found");
@@ -59,11 +76,18 @@ public class BookController {
     }
 
     @DeleteMapping("/books/{id}")
-    public ResponseEntity<Object> deleteBook(@PathVariable(value = "id") UUID id) {
+    public ResponseEntity<Object> deleteBook(@PathVariable(value = "id") UUID id, HttpServletRequest request) {
         Optional<Book> book = bookRepository.findById(id);
+        var user = userRepository.findById((UUID) request.getAttribute("idUser"));
+
+        if(user.get().getAdmin().equals(false)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not allowed");
+        }
+
         if(book.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Book not found");
         }
+
         bookRepository.delete(book.get());
         return ResponseEntity.status(HttpStatus.OK).body("Book deleted successfully");
     }
